Move applePayProcessor to own file
