+++ client/my-sites/checkout/composite-checkout/payment-method-processors.js
@@ -1,0 +1,34 @@
+/**
+ * External dependencies
+ */
+import { defaultRegistry } from '@automattic/composite-checkout';
+
+/**
+ * Internal dependencies
+ */
+import {
+	getDomainDetails,
+	wpcomTransaction,
+	submitApplePayPayment,
+} from './payment-method-helpers';
+
+const { select, dispatch } = defaultRegistry;
+
+export function applePayProcessor( submitData ) {
+	const pending = submitApplePayPayment(
+		{
+			...submitData,
+			siteId: select( 'wpcom' )?.getSiteId?.(),
+			domainDetails: getDomainDetails( select ),
+			country: select( 'wpcom' )?.getContactInfo?.()?.countryCode?.value,
+			postalCode: select( 'wpcom' )?.getContactInfo?.()?.postalCode?.value,
+		},
+		wpcomTransaction
+	);
+	// save result so we can get receipt_id and failed_purchases in getThankYouPageUrl
+	pending.then( ( result ) => {
+		// TODO: do this automatically when calling setTransactionComplete
+		dispatch( 'wpcom' ).setTransactionResponse( result );
+	} );
+	return pending;
+}
